refactor: move models to domain
